requirements.txt
streamlit
requests
plotly
streamlit_shadcn_ui
python-dotenv